requirements.txt
flask
dotenv
google-genai
